refactor: restructure package layout
