+++ src/main/java/com/devcard/devcard/card/controller/CardController.java
@@ -1,9 +1,9 @@
-package com.devcard.devcard.controller;
+package com.devcard.devcard.card.controller;
 
-import com.devcard.devcard.dto.CardRequestDto;
-import com.devcard.devcard.dto.CardResponseDto;
-import com.devcard.devcard.vo.Card;
-import com.devcard.devcard.service.CardService;
+import com.devcard.devcard.card.dto.CardResponseDto;
+import com.devcard.devcard.card.vo.Card;
+import com.devcard.devcard.card.dto.CardRequestDto;
+import com.devcard.devcard.card.service.CardService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
