+++ src/main/java/com/devcard/devcard/card/service/CardService.java
@@ -1,7 +1,7 @@
-package com.devcard.devcard.service;
+package com.devcard.devcard.card.service;
 
-import com.devcard.devcard.vo.Card;
-import com.devcard.devcard.repository.CardRepository;
+import com.devcard.devcard.card.repository.CardRepository;
+import com.devcard.devcard.card.vo.Card;
 import org.springframework.stereotype.Service;
 
 @Service
